backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
pyodbc==5.0.1
azure-storage-blob==12.19.0
azure-identity==1.15.0
Pillow==10.2.0

